TODO: ipcache: Support incremental ipcache updates
Previously, the InjectLabels() function which reacts to updates to the
IPCache metadata would iterate through the entire metadata map in order
to detect changes that occur, and subsequently react to those updates.
The reaction includes potentially allocating new identities, then
triggering updates into the policymaps and the datapath's ipcache.
Given that there was no way to track which prefixes had been updated,
this meant iterating over the entire map, perhaps re-allocating
identities for prefixes that had not changed. This also made it more
difficult in future to support different types of updates to the ipcache
from other sources, since each feature would need to have special logic
encoded in this function to correctly handle the identity allocation.

Rework the InjectLabels() implementation to handle a queue of prefixes
to update instead.

TODO: This can currently leak identities if someone does an 'add' and
then another subsequent 'add'. The logic here needs to be combined with
the 'update' logic in the subsequent patch to ensure that in this
situation, any references held for a prior identity are released when
the identity is updated.

Co-authored-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>