datapath: Allow egress GW with XDP
This commit allows a user to use the egress GW feature with XDP
(--bpf-lb-acceleration != disabled). However, a support for the
redirection from the reverse egress GW traffic to a tunnel netdev from
bpf_xdp is still missing [1].

With the current approach, when enabling the egress GW together with XDP
a reply packets might be dropped if the iptables' default policy for the
FORWARD chain is DROP.

The egress GW reply packet will be handled by the bpf_xdp which will do
the reverse SNAT translation. Also, it will set XFER_PKT_NO_SVC to
prevent the bpf_host (attached to the same netdev as the bpf_xdp) to
call nodeport_lb4() (perf optimization). The latter function is the
entry point not only for the NodePort BPF, but also for the egress GW.
Therefore, due to XFER_PKT_NO_SVC, the reply won't be handled by the
bpf_host. As a consequence, it won't be redirected to the tunnel netdev
and it will be passed to the stack. Instead of redirecting the reply to
the tunnel netdev, the stack might drop the packet if the FORWARD policy
is set to DROP.

This will be fixed in [1] once we do the encap from bpf_xdp.

[1]: https://github.com/cilium/cilium/issues/17770

Signed-off-by: Martynas Pumputis <m@lambda.lt>